pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "parkinsons-voice"
version = "0.1.0"
description = "Parkinson's Disease Detection from Voice using Wav2Vec2"
authors = [{name = "Parkinsons Voice Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
]
dependencies = [
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "transformers>=4.34.0",
    "datasets>=2.14.5",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "scikit-learn>=1.3.0",
    "librosa>=0.10.1",
    "scipy>=1.11.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.1",
    "jupyter>=1.0.0",
]
viz = [
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
]
experiment = [
    "wandb>=0.21.3",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"